Refactor dev scripts for professionalism and consistency
This commit improves the development scripts and database initialization process to provide a cleaner and more professional developer experience.

Key changes:
- Renamed SQL files to a numbered sequence (`01_schema.sql`, `02_views.sql`, `03_reset.sql`) to make the execution order explicit.
- Renamed helper scripts (`db_runner.py` -> `db_cli.py`, `procedural_seed.py` -> `db_seeder.py`) for better clarity.
- Unified the database initialization logic in `dev.sh` and `dev.ps1` to use a single Python script (`db_cli.py`).
- This eliminates direct `psql` calls from the PowerShell script, which resolves the issue of verbose, localized log messages and ensures a consistent, clean output across all platforms.
